Do not cache NULL parser in SourceCache if parserFactory is not (yet) present
It was observed that when javascript2.editor is initialized via
Feature-On-Demand, all features for this file, that rely on the parser,
don't work correctly. Debugging showed, that for these files the
parser was NULL, but parserInitialized was true and thus no second try
to get a working parser was made.